#15 return requeueable errors from upgrade manager
The component CR status is used to hold the state of the respective
time when the CR should be requeued. The interval doubles until a
threshold is reached; then only a fixed interval will be used.

The initial value of 15 seconds was plain guessed, so there was no deep
thinking about that value.

Co-authored-by: Jeremias Weber <jeremias.weber@cloudogu.com>