#29 Add RBAC
Add permissions for longhorn-system namespace and volumesnapshotclasses for backup provider installation.

Attention: With this change the target `k8s-apply` throws an error because of different namespaces. We should consider to delete the namespace option in the target and rely on the actual configured namespace or ignore this because the `helm-apply` works fine and will be used in future instead of `k8s-apply`