#51 Reset status of CRs
Requeue CRs on failures. This resets the status so that the CRs can be modified and reconciled again.
Add 3 Additional status phases for installing, upgrading, and deleting to identify the state of a CR in reconcile mode. This should be used determining the health of a component. In reconcile with new status phases the component is not healthy.